Auto stash before rebase of "refactor-to-new-structure"
