+++ .kiro/specs/handler-based-architecture/tasks.md
@@ -96,16 +96,16 @@
   - グローバルエラー処理ロジックを実装
   - _要件: 1.9, 3.1, 3.2, 3.3_
 
-- [ ] 5. 型定義の再編成
+- [x] 5. 型定義の再編成
   - 共通型定義を整理し、ハンドラー固有の型を分離する
   - _要件: 3.3, 6.1, 6.2_
 
-- [ ] 5.1 共通型定義の整理
+- [x] 5.1 共通型定義の整理
   - `src/types/index.ts`からハンドラー固有の型を削除
   - 共通型のみを残して統合エクスポートを更新
   - _要件: 3.3, 6.1_
 
-- [ ] 5.2 ハンドラーレジストリの更新
+- [x] 5.2 ハンドラーレジストリの更新
   - `src/types/handler-registry.ts`を新しいハンドラー構造に対応
   - 必要に応じてレジストリ定義を更新
   - _要件: 4.5, 6.1_
