+++ src/handlers/command/parser.ts
@@ -12,13 +12,16 @@ import {
   isObjectLiteralNode,
 } from '../../utils/guards/ast.js';
 import type {
-  CommandDefinition,
   CommandParser,
+  ParsedCommandDefinition,
   ValidationResult,
 } from './types.js';
 
 export class CommandParserImpl implements CommandParser {
-  async parse(content: string, filePath: string): Promise<CommandDefinition> {
+  async parse(
+    content: string,
+    filePath: string
+  ): Promise<ParsedCommandDefinition> {
     const sourceFile = ts.createSourceFile(
       filePath,
       content,
@@ -35,7 +38,7 @@ export class CommandParserImpl implements CommandParser {
     }
 
     const description = this.extractDescription(sourceFile);
-    const definition: CommandDefinition = {
+    const definition: ParsedCommandDefinition = {
       name: commandName,
       path: filePath,
       hasParams: false, // Will be updated by scanner
@@ -50,7 +53,7 @@ export class CommandParserImpl implements CommandParser {
     return definition;
   }
 
-  validate(definition: CommandDefinition): ValidationResult {
+  validate(definition: ParsedCommandDefinition): ValidationResult {
     const errors: string[] = [];
 
     if (!definition.name) {
@@ -143,9 +146,9 @@ export class CommandParserImpl implements CommandParser {
 
 export async function parseFiles(
   files: CommandFile[]
-): Promise<CommandDefinition[]> {
+): Promise<ParsedCommandDefinition[]> {
   const parser = new CommandParserImpl();
-  const definitions: CommandDefinition[] = [];
+  const definitions: ParsedCommandDefinition[] = [];
 
   // 並列処理で高速化
   const parsePromises = files.map(async (file) => {
