+++ src/handlers/global-error/index.ts
@@ -1,5 +1,4 @@
 import type {
-  CLIError,
   ErrorDisplayOptions,
   GlobalErrorContext,
   GlobalErrorExecutionOptions,
