+++ src/handlers/help/types.ts
@@ -1,5 +1,27 @@
-import type { HelpContext } from '../../types/context.js';
-import type { HelpHandler, HelpHandlerFactory } from '../../types/handlers.js';
+import type { Context } from '../../types/context.js';
+
+// ヘルプハンドラー固有の型定義をここに移動
+
+/**
+ * ヘルプハンドラー
+ */
+export interface HelpHandler {
+  /** コマンド名 */
+  name?: string;
+  /** 説明 */
+  description?: string;
+  /** 使用例 */
+  examples?: string[];
+  /** エイリアス */
+  aliases?: string[];
+  /** 追加のヘルプテキスト */
+  additionalHelp?: string;
+}
+
+/**
+ * ヘルプハンドラー用のコンテキスト
+ */
+export type HelpContext<E = typeof process.env> = Context<E>;
 
 /**
  * ヘルプハンドラー関連の型定義
@@ -8,7 +30,10 @@ import type { HelpHandler, HelpHandlerFactory } from '../../types/handlers.js';
 /**
  * ヘルプハンドラーのファクトリー関数型
  */
-export type { HelpHandlerFactory, HelpHandler, HelpContext };
+export interface HelpHandlerFactory<E = typeof process.env> {
+  (context: HelpContext<E>): HelpHandler;
+  (): HelpHandler;
+}
 
 /**
  * ヘルプ情報の処理結果
