+++ src/types/index.ts
@@ -1,32 +1,17 @@
-// 型定義の統合エクスポート
+// 共通型定義の統合エクスポート
 
-// 実行コンテキスト関連
+// 実行コンテキスト関連（共通）
 export type {
   BaseCommandContext,
   CommandContext,
   CommandHandler,
   Context,
-  EnvContext,
-  ErrorContext,
-  GlobalErrorContext,
-  HelpContext,
-  MiddlewareFactoryContext,
-  ParamsContext,
-  VersionContext,
 } from './context.js';
 
-// コマンド定義関連
-export type {
-  CommandDefinition,
-  CommandDefinitionFactory,
-  DynamicParam,
-  ParsedCommand,
-} from './definition.js';
-// エラー型定義
+// エラー型定義（共通）
 export type {
   CLIError,
   ModuleError,
-  ValidationError as CLIValidationError,
   ValidationError,
   ValidationIssue,
 } from './errors.js';
@@ -36,12 +21,13 @@ export {
   isModuleError,
   isValidationError,
 } from './errors.js';
+
+// ハンドラーレジストリ（共通）
 export type {
   HandlerDefinition,
   HandlerRegistryMap,
   HandlerScope,
 } from './handler-registry.js';
-// ハンドラーレジストリ
 export {
   createHandlerRegistryMap,
   EXECUTION_ORDER,
@@ -50,41 +36,9 @@ export {
   HANDLER_REGISTRY,
   validateHandlerDependencies,
 } from './handler-registry.js';
-// ハンドラー関連
+
+// バリデーション関連（共通）
 export type {
-  EnvHandler,
-  GlobalErrorHandler,
-  HelpHandler,
-  VersionHandler,
-} from './handlers.js';
-// メタデータ関連
-export type { CommandMetadata } from './metadata.js';
-// ミドルウェア関連
-export type {
-  MiddlewareContext,
-  MiddlewareExport,
-  MiddlewareFactory,
-  MiddlewareHandler,
-  NextFunction,
-} from './middleware.js';
-// バリデーション関連（valibotのみ）
-export type {
-  CommandHandlerFactory,
-  EnvDefinitionFunction,
-  EnvHandlerFactory,
-  ErrorHandler,
-  ErrorHandlerFactory,
-  GlobalErrorHandlerFactory,
-  HelpHandlerFactory,
-  MiddlewareHandlerFactory,
-  ParamMapping,
-  ParamsDefinitionFunction,
-  ParamsHandler,
-  // Factory types with inference
-  ParamsHandlerFactory,
   ValidationFunction,
   ValidationResult,
-  VersionDefinitionFunction,
-  VersionHandlerFactory,
 } from './validation.js';
-export { SCHEMA_TYPE } from './validation.js';
