+++ src/utils/validation/index.ts
@@ -1,13 +1,17 @@
 import type * as v from 'valibot';
-import type { ValidationError, ValidationIssue } from '../../types/errors.js';
 import type {
   EnvFieldSchema,
   EnvSchema,
   EnvValidationResult,
+} from '../../handlers/env/types.js';
+import type {
   ManualFieldSchema,
   ManualSchema,
   ParamMapping,
   ParamsHandler,
+} from '../../handlers/params/types.js';
+import type { ValidationError } from '../../types/errors.js';
+import type {
   ValidationFunction,
   ValidationResult,
 } from '../../types/validation.js';
